fix: market sell item bug (#379)
This PR fixes a bug with Market where:

Player 1 creates an offer for an item and goes hunting.
Player 2 goes to the Market to look for something to buy.
Player 1 dies and stays dead on the game's death screen.
Player 2 buys the item that Player 1 had put up for sale.
Player 1 logs in after Player 2 buys the item.

Player 2 received the item they bought that belonged to Player 1 in the Market.
Player 1 did not receive the money for the item they sold to Player 2.